service/settings: Fix job polling mechanism
Instead of using the unauthorized internal job URL, which is returned by the API, parse the job uuid from the URL and use that ID for polling on the proper endpoint.

Ideally this is fixed in the API, but this is how it's been done in kamikaze so far as well, so let's first start here.